ai-city-views-finder/requirements_cloudrun.txt
# Requirements for Cloud Run deployment with Streamlit

# Streamlit
streamlit==1.40.0

# Google Cloud Platform
google-cloud-aiplatform[adk,agent_engines]==1.132.0
google-cloud-logging==3.11.4

# Core dependencies
pydantic==2.12.1
cloudpickle==3.1.1

# Additional utilities
python-dotenv==1.0.0
nest-asyncio==1.6.0
